code/requirements.txt
azure-ai-projects>=2.0.0b1
azure-identity
python-dotenv
Flask>=3.1.0
Markdown>=3.7
bleach>=6.2.0